feat,refactor: Add debug option, pring json, move post data
Added debug option to remove debugging statements and writes so
that the IO can be used to communicate to a backend. When in
no-debug mode (with no-d in argv) the program only prints the json
of the links related to the post.
Some refactoring of the headers and data for the post call